Trapi updates (#122)
* Backplane overhaul to enable centrizied of configuration and conversion between trapi versions

* Backplane overhaul to enable centrizied of configuration and conversion between trapi versions

* Addition of roger, update tranql frontend to talk to backplane always to ensure trapi configs are universal.
Updating tests to those changes

* docker image version bump

* gnbr and indigo

* Tests were that were using local server  to use mock servers

* icees update to latest bl lookup changes

* update tests to match biolink:PascalConceptType in tests
