+++ tranql/tests/mock/automat-registry.json
@@ -1,0 +1,4 @@
+[
+  "intact",
+  "viral-proteome"
+]
